test(client): fold backend compat coverage
