+++ env.d.ts
@@ -1,7 +1,6 @@
 declare namespace NodeJS {
   interface ProcessEnv {
     CI: string | undefined;
-    POLYMARKET_RUN_BACKEND_COMPAT_TESTS: string | undefined;
     POLYMARKET_RUN_METERED_TESTS: string | undefined;
     POLYMARKET_DEPOSIT_WALLET: string | undefined;
     POLYMARKET_PRIVATE_KEY: string | undefined;
