+++ package.json
@@ -20,7 +20,6 @@
     "release:canary": "changeset publish --tag latest --no-git-tag",
     "release:latest": "node scripts/update-latest-dist-tags.mjs",
     "test:bindings": "vitest --project bindings --passWithNoTests",
-    "test:backend-compat": "POLYMARKET_RUN_BACKEND_COMPAT_TESTS=1 vitest --project backend-compat",
     "test:client": "vitest --project client",
     "test:client:integration": "vitest --project client-integration",
     "test:types": "vitest --project types --passWithNoTests",
