+++ packages/client/tests/integration/activity.test.ts
@@ -1,5 +1,6 @@
 import { expectPresent } from '@polymarket/types';
 import { describe, expect, it } from './fixtures';
+import { expectNonEmptyPage, expectPageWindow } from './helpers';
 
 const TEST_USER = '0x7c3db723f1d4d8cb9c550095203b686cb11e5c6b';
 
@@ -21,22 +22,33 @@ describe('Activity', () => {
         }),
       );
     });
+
+    it('lists global trades across pages', async ({ publicClient }) => {
+      const paginator = publicClient.listTrades({
+        pageSize: 100,
+      });
+      const firstPage = await paginator.firstPage().then(expectNonEmptyPage);
+
+      expect(firstPage.items.length).toBeGreaterThan(0);
+      await expectPageWindow(paginator, firstPage, 29);
+    });
   });
 
   describe('listActivity', () => {
     it('lists wallet activity', async ({ publicClient }) => {
-      const result = await publicClient
-        .listActivity({
-          user: TEST_USER,
-          pageSize: 1,
-        })
-        .firstPage();
+      const paginator = publicClient.listActivity({
+        user: TEST_USER,
+        pageSize: 100,
+        type: ['TRADE'],
+      });
+      const result = await paginator.firstPage().then(expectNonEmptyPage);
 
-      expect(result.items).toHaveLength(1);
+      expect(result.items.length).toBeGreaterThan(0);
+      await expectPageWindow(paginator, result, 29);
       expect(expectPresent(result.items[0])).toEqual(
         expect.objectContaining({
           conditionId: expect.any(String),
-          type: expect.any(String),
+          type: 'TRADE',
           wallet: TEST_USER,
         }),
       );
