+++ vitest.config.ts
@@ -35,10 +35,7 @@ export default defineConfig({
         test: {
           name: 'client',
           include: ['packages/client/src/**/*.test.ts'],
-          exclude: [
-            ...configDefaults.exclude,
-            'packages/client/src/backend-compat.test.ts',
-          ],
+          exclude: [...configDefaults.exclude],
           environment: 'node',
           testTimeout: 10_000,
           typecheck: {
@@ -56,16 +53,6 @@ export default defineConfig({
           exclude: [...configDefaults.exclude],
           environment: 'node',
           testTimeout: 60_000,
-        },
-      },
-      {
-        extends: true,
-        test: {
-          name: 'backend-compat',
-          include: ['packages/client/src/backend-compat.test.ts'],
-          exclude: [...configDefaults.exclude],
-          environment: 'node',
-          testTimeout: 600_000,
         },
       },
     ],
